include module in route if applicable
The property_exists prevents the route from including the module if module exists. CWebModule would be the first approach, but hence messing up with the framework is very common the safest bet is CModule. If CModule instance is returned then it is dead certain that getId() exists and thus is safe to use.